move inspectDeal to cli package
